Updated script

+++ azure/powershell/messaging/event-hubs/Create-EventHub.ps1
@@ -66,10 +66,10 @@ $eventHubNamespace
 # --------------- 4 --------------- 
 Write-Host "---> Creating an event hub in the namespace you created" -ForegroundColor Green
 $rndeventhub = (New-Guid).ToString().Split("-")[0]
-$env:eventhubName = "" # Initialization - With PowerShell's StrictMode set to ON uninitialized variables are flagged as an error.
-$env:eventhubName = "test_eventhub_$rndeventhub"
+$env:varEventHubName = "" # Initialization - With PowerShell's StrictMode set to ON uninitialized variables are flagged as an error.
+$env:varEventHubName = "test_eventhub_$rndeventhub"
 $retentionInDays = 1  # Message Retention customization is not available in a Basic Tier Namespace. Change variable $paramNamespaceSku to "Standard" to increase message retention.
-$eventhub = New-AzEventHub -ResourceGroupName "$paramResourceGroup" -NamespaceName "$paramEventHubNamespace" -Name "$env:eventhubName" -MessageRetentionInDays $retentionInDays
+$eventhub = New-AzEventHub -ResourceGroupName "$paramResourceGroup" -NamespaceName "$paramEventHubNamespace" -Name "$env:varEventHubName" -MessageRetentionInDays $retentionInDays
 Write-Host "---> Event Hub details:" -ForegroundColor Green
 $eventhub
 
@@ -78,30 +78,39 @@ $eventhub
 Write-Host "---> Get the primary connection string for the namespace" -ForegroundColor Green
 $eventhubKey = Get-AzEventHubKey -ResourceGroupName "$paramResourceGroup" -Namespace "$paramEventHubNamespace" -Name "RootManageSharedAccessKey"
 $eventhubKey
-$env:ehubNamespaceConnectionString = "" # Initialization - With PowerShell's StrictMode set to ON uninitialized variables are flagged as an error.
-$env:ehubNamespaceConnectionString = $eventhubKey.PrimaryConnectionString
+$env:varEventHubNamespaceConnection = "" # Initialization - With PowerShell's StrictMode set to ON uninitialized variables are flagged as an error.
+$env:varEventHubNamespaceConnection = $eventhubKey.PrimaryConnectionString
 Write-Host "---> Primary Connection String" -ForegroundColor Green
-$env:ehubNamespaceConnectionString
+$env:varEventHubNamespaceConnection
 
 
 # --------------- 6 --------------- 
 Write-Host "---> Create an application to Send events" -ForegroundColor Green
-# https://docs.microsoft.com/en-us/azure/event-hubs/event-hubs-dotnet-standard-getstarted-send
-$appFolderName = "EventHubsAppSend"
-$appProgramFile = "ProgramEH-Send.cs"
-if ( Test-Path -Path $appFolderName -PathType Container ) { Remove-Item -path $appFolderName -Recurse –force }
-dotnet new console -n $appFolderName
-Set-Location $appFolderName
-dotnet add package Azure.Messaging.EventHubs
-Copy-Item ../$appProgramFile .
-Move-Item $appProgramFile Program.cs -Force
-Write-Host "---> Check your results" -ForegroundColor Green
-dotnet build
-# NOTE: If you receive a Timeout Exception, your environment is probably blocking the TCP ports
-# used for AMQP by default. You need to have ports 5671 and 5672 opened for AMQP connections.
-dotnet run
+
+$runtimeStack = "JS"
+if ($runtimeStack -eq "C#") {
+  # https://docs.microsoft.com/en-us/azure/event-hubs/event-hubs-dotnet-standard-getstarted-send
+  $appFolderName = "EventHubsAppSend"
+  $appProgramFile = "ProgramEH-Send.cs"
+  if ( Test-Path -Path $appFolderName -PathType Container ) { Remove-Item -path $appFolderName -Recurse –force }
+  dotnet new console -n $appFolderName
+  Set-Location $appFolderName
+  dotnet add package Azure.Messaging.EventHubs
+  Copy-Item ../$appProgramFile .
+  Move-Item $appProgramFile Program.cs -Force
+  Write-Host "---> Check your results" -ForegroundColor Green
+  dotnet build
+  # NOTE: If you receive a Timeout Exception, your environment is probably blocking the TCP ports
+  # used for AMQP by default. You need to have ports 5671 and 5672 opened for AMQP connections.
+  dotnet run
+  Set-Location ..
+} 
+elseif ($runtimeStack -eq "JS") {
+  npm install @azure/storage-blob
+  npm install @azure/eventhubs-checkpointstore-blob
+  node send.js
+}
 Write-Host "---> In the Azure portal, you can verify that the event hub has received the messages. Switch to Messages view in the Metrics section." -ForegroundColor Green
-Set-Location ..
 
 
 # --------------- 7 --------------- 
@@ -132,34 +141,40 @@ $storageAccount
 Write-Host "---> Get storage account key and connection string" -ForegroundColor Green
 $accountKey = Get-AzStorageAccountKey -ResourceGroupName "$paramResourceGroup" -AccountName "$paramStorageAccount" |
 Where-Object { $_.KeyName -eq "Key1" } | Select-Object -ExpandProperty Value
-$env:storageConnectionString = "" # Initialization - With PowerShell's StrictMode set to ON uninitialized variables are flagged as an error.
-$env:storageConnectionString = "DefaultEndpointsProtocol=https;AccountName=$paramStorageAccount;AccountKey=$accountKey"
-$env:storageConnectionString
+$env:varStorageConnectionString = "" # Initialization - With PowerShell's StrictMode set to ON uninitialized variables are flagged as an error.
+$env:varStorageConnectionString = "DefaultEndpointsProtocol=https;AccountName=$paramStorageAccount;AccountKey=$accountKey"
+$env:varStorageConnectionString
 
 
 # --------------- 9 --------------- 
 Write-Host "---> Create a blob container" -ForegroundColor Green
 $ctx = $storageAccount.Context
-$env:containerName = "" # Initialization - With PowerShell's StrictMode set to ON uninitialized variables are flagged as an error.
-$env:containerName = "messages"
-New-AzStorageContainer -Name "$env:containerName" -Context $ctx -Permission "Blob"
+$env:varContainerName = "" # Initialization - With PowerShell's StrictMode set to ON uninitialized variables are flagged as an error.
+$env:varContainerName = "messages"
+New-AzStorageContainer -Name "$env:varContainerName" -Context $ctx -Permission "Blob"
 
 
 # --------------- 10 --------------- 
 Write-Host "---> Create an application to Receive events" -ForegroundColor Green
-# https://docs.microsoft.com/en-us/azure/event-hubs/event-hubs-dotnet-standard-getstarted-send
-$appFolderName = "EventHubsAppReceive"
-$appProgramFile = "ProgramEH-Receive.cs"
-if ( Test-Path -Path $appFolderName -PathType Container ) { Remove-Item -path $appFolderName -Recurse –force }
-dotnet new console -n $appFolderName
-Set-Location $appFolderName
-dotnet add package Azure.Messaging.EventHubs
-dotnet add package Azure.Messaging.EventHubs.Processor
-Copy-Item ../$appProgramFile .
-Move-Item $appProgramFile Program.cs -Force
-Write-Host "---> Check your results" -ForegroundColor Green
-dotnet build
-# NOTE: If you receive a Timeout Exception, your environment is probably blocking the TCP ports
-# used for AMQP by default. You need to have ports 5671 and 5672 opened for AMQP connections.
-dotnet run
-Set-Location ..
+
+if ($runtimeStack -eq "C#") {
+  # https://docs.microsoft.com/en-us/azure/event-hubs/event-hubs-dotnet-standard-getstarted-send
+  $appFolderName = "EventHubsAppReceive"
+  $appProgramFile = "ProgramEH-Receive.cs"
+  if ( Test-Path -Path $appFolderName -PathType Container ) { Remove-Item -path $appFolderName -Recurse –force }
+  dotnet new console -n $appFolderName
+  Set-Location $appFolderName
+  dotnet add package Azure.Messaging.EventHubs
+  dotnet add package Azure.Messaging.EventHubs.Processor
+  Copy-Item ../$appProgramFile .
+  Move-Item $appProgramFile Program.cs -Force
+  Write-Host "---> Check your results" -ForegroundColor Green
+  dotnet build
+  # NOTE: If you receive a Timeout Exception, your environment is probably blocking the TCP ports
+  # used for AMQP by default. You need to have ports 5671 and 5672 opened for AMQP connections.
+  dotnet run
+  Set-Location ..
+}
+elseif ($runtimeStack -eq "JS") {
+  node receive.js
+}
